Change Database structure
