--- backend/Database/create_db.py
@@ -1,11 +1,0 @@
-from backend.backend import app, db
-
-with app.app_context():
-    # create the database and the db table
-    db.create_all()
-
-    # commit the changes
-    db.session.commit()
-
-
-
