+++ backend/backend.py
@@ -1,6 +1,6 @@
 from app import create_app
 from flask_cors import CORS
-from Database import config
+from routes.Database import config
 
 app = create_app()
 # app = Flask(__name__)
@@ -9,7 +9,7 @@ app.config['SQLALCHEMY_DATABASE_URI'] = config.SQLALCHEMY_DATABASE_URI_1
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = config.SQLALCHEMY_TRACK_MODIFICATIONS_1
 
 # initialize DB
-from routes.Models.main_system_model import t_event_filter_db
+from routes.Database.Event_DB import t_event_filter_db
 t_event_filter_db.init_app(app)
 
 with app.app_context():
